Revert "Add: Modal to choose a start pattern on new templates. (#46248)"
This reverts commit ad9624ab8970a54862f7d897190e5ba84b64b231.
