+++ lib/compat/wordpress-6.1/rest-api.php
@@ -6,6 +6,22 @@
  */
 
 /**
+ * Update `wp_template` and `wp_template-part` post types to use
+ * Gutenberg's REST controller.
+ *
+ * @param array  $args Array of arguments for registering a post type.
+ * @param string $post_type Post type key.
+ */
+function gutenberg_update_templates_template_parts_rest_controller( $args, $post_type ) {
+	if ( in_array( $post_type, array( 'wp_template', 'wp_template-part' ), true ) ) {
+		$args['rest_controller_class'] = 'Gutenberg_REST_Templates_Controller';
+	}
+	return $args;
+}
+add_filter( 'register_post_type_args', 'gutenberg_update_templates_template_parts_rest_controller', 10, 2 );
+
+
+/**
  * Add the post type's `icon`(menu_icon) in the response.
  * When we backport this change we will need to add the
  * `icon` to WP_REST_Post_Types_Controller schema.
