+++ lib/compat/wordpress-6.2/rest-api.php
@@ -6,21 +6,6 @@
  */
 
 /**
- * Update `wp_template` and `wp_template-part` post types to use
- * Gutenberg's REST controller.
- *
- * @param array  $args Array of arguments for registering a post type.
- * @param string $post_type Post type key.
- */
-function gutenberg_update_templates_template_parts_rest_controller( $args, $post_type ) {
-	if ( in_array( $post_type, array( 'wp_template', 'wp_template-part' ), true ) ) {
-		$args['rest_controller_class'] = 'Gutenberg_REST_Templates_Controller_6_2';
-	}
-	return $args;
-}
-add_filter( 'register_post_type_args', 'gutenberg_update_templates_template_parts_rest_controller', 10, 2 );
-
-/**
  * Registers the block pattern categories REST API routes.
  */
 function gutenberg_register_rest_block_pattern_categories() {
